# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API for managing mechanics, customers, inventory, and service tickets."
  version: "1.0.0"
host: "127.0.0.1:5000" # render name: "your-app-name.onrender.com"
schemes:
  - "https"
basePath: "/"
securityDefinitions:
  BearerAuth:
    type: apiKey
    swagger: "2.0"
    in: header
    description: "JWT token using Bearer token number. Example: 'Bearer {token number}'"
paths:
  /inventory/mechanic/:
    get:
      tags: [Inventory]
      summary: "Mechanic gets paginated list of parts with pricing"
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
        - name: per_page
          in: query
          type: integer
      responses:
        200:
          description: "List of parts with pricing"
        400:
          description: "Bad request"
        500:
          description: "Internal server error"
  /inventory/mechanic/{id}:
    get:
      tags: [Inventory]
      summary: "Mechanic gets part by id with pricing"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Part details with pricing"
        404:
          description: "Not found"
        500:
          description: "Internal server error"
...
  /service-tickets/{ticket_id}/mechanics:
    get:
      tags:
        - ServiceTicket
      summary: Mechanic gets mechanics for ticket
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of mechanics
        '500':
          description: Server error
...
  /service-tickets/mechanic/my-tickets:
    get:
      tags:
        - ServiceTicket
      summary: Mechanic gets own assigned tickets (paginated, status filter)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of assigned tickets
        '500':
          description: Server error
...
  /mechanics/login:
    post:
      tags: [Mechanics]
      summary: "Mechanic login"
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            $ref: "#/definitions/MechanicLogin"
      responses:
        200:
          description: "Login successful (JWT token returned)"
        401:
          description: "Invalid credentials"
        500:
          description: "Internal server error"
  /mechanics/my-tickets:
    get:
      tags: [Mechanics]
      summary: "Get my assigned tickets"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "List of tickets assigned to mechanic"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /mechanics/{id}:
    get:
      tags: [Mechanics]
      summary: "Get mechanic by ID"
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic details"
        404:
          description: "Mechanic not found"
        500:
          description: "Internal server error"
    put:
      tags: [Mechanics]
      summary: "Update mechanic profile"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: mechanic
          required: true
          schema:
            $ref: "#/definitions/MechanicUpdate"
      responses:
        200:
          description: "Mechanic updated"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Mechanic not found"
        500:
          description: "Internal server error"
    delete:
      tags: [Mechanics]
      summary: "Delete mechanic (admin only)"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic deleted"
        404:
          description: "Mechanic not found"
        500:
          description: "Internal server error"
  /mechanics/profile:
    get:
      tags: [Mechanics]
      summary: "Get own mechanic profile"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Mechanic profile"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /mechanics/dashboard:
    get:
      tags: [Mechanics]
      summary: "Get mechanic dashboard information"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Dashboard data with tickets & stats"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /mechanics/secure-data:
    get:
      tags: [Mechanics]
      summary: "Access secure data"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Protected mechanic information"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /mechanics/change-password:
    put:
      tags: [Mechanics]
      summary: "Change mechanic password"
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: passwords
          required: true
          schema:
            $ref: "#/definitions/PasswordChange"
      responses:
        200:
          description: "Password successfully changed"
        401:
          description: "Old password is incorrect"
        500:
          description: "Internal server error"
  /mechanics/admin/update/{id}:
    put:
      tags: [Mechanics]
      summary: "Admin update mechanic"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: mechanic
          required: true
          schema:
            $ref: "#/definitions/MechanicUpdate"
      responses:
        200:
          description: "Mechanic updated by admin"
        404:
          description: "Mechanic not found"
        500:
          description: "Internal server error"
  /mechanics/admin/create:
    post:
      tags: [Mechanics]
      summary: "Create admin"
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: admin
          required: true
          schema:
            $ref: "#/definitions/AdminCreate"
      responses:
        201:
          description: "Admin created"
        409:
          description: "Admin already exists"
        500:
          description: "Internal server error"
  /mechanics/admin/login:
    post:
      tags: [Mechanics]
      summary: "Admin login"
      parameters:
        - in: body
          name: adminLogin
          required: true
          schema:
            $ref: "#/definitions/AdminLogin"
      responses:
        200:
          description: "Admin login successful"
        401:
          description: "Invalid admin credentials"
        500:
          description: "Internal server error"
...
definitions:
...
  MechanicLogin:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
  MechanicUpdate:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      address:
        type: string
  PasswordChange:
    type: object
    properties:
      old_password:
        type: string
      new_password:
        type: string
    required:
      - old_password
      - new_password
  AdminCreate:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    required:
      - username
      - password
  AdminLogin:
    type: object
    properties:
      username:
        type: string
        example: "admin"
      password:
        type: string
        example: "admin123"
    required:
      - username
      - password
...